closes https://github.com/kimkulling/osre/issues/67: rework SDL2-threading API.
